feat: add support for attaching images in LLMAssistant
This commit adds support for attaching images in the `LLMAssistant` class. It introduces a new parameter `attached_images` in the `run_prompt` method, allowing users to pass a list of image file paths to be attached to the user input. The attached images are encoded as base64 and sent to the OpenAI API as part of the user input.